Created CMPT395_Assignment_1_Kata_3.py and Completed First Test of Kata 3
Created the password_Validator function which takes a string acting as a password and passes it through just one validation check for now. The check is for its lenght with the minimum length allowed being 8 characters long. The function then passes a boolean for if it fails or validates, and an error message for what it failed to validate. The first test for the function checks what happens if password that is too short is passed into it, and what happens if a long enough password is passed into it.